Simplified spatial dataset of country boundaries.

SimpleCountries is a simplified world borders dataset suitable for global maps and quick spatial searches. This dataset is distributed with the package and is can be used with getCountry(), getCountryCode() and getCountryName() when restricting searches to land-based locations.

SimpleCountries

Format

A simple features data frame with 246 records and 7 columns of data.

Details

This dataset is equivalent to TMWorldBordersSimple but with fewer columns of data.

See also