Simplified spatial dataset of world timezones.

This dataset is used by default in the getTimezones() function and contains the following columns of data in the @data slot:

timezone -- Olson timezone

UTC_offset -- offset from UTC (hours)

UTC_DST_offset -- offset from UTC during daylight savings (hours)

countryCode -- ISO 3166-1 alpha-2 country code

longitude -- longitude of the timezone polygon centroid

latitude -- longitude of the timezone polygon centroid

status -- one of 'Canonical', 'Alias' or 'Deprecated'

notes -- typically specifying the target of an 'Alias'

polygonID -- unique identifier (= timezone)

This dataset was generated on 2020-11-12 by running:

library(MazamaSpatialUtils)
setSpatialDataDir("~/Data/Spatial")

convertWorldTimezones()

loadSpatialData("WorldTimezones_02")

SimpleTimezones <- WorldTimezones_02
save(SimpleTimezones, file = "data/SimpleTimezones.rda")

SimpleTimezones

Format

A SpatialPolygonsDataFrame with 423 records and 9 columns of data.