Convert (Simple) World Borders Shapefile

Returns a SpatialPolygonsDataFrame for simple world divisions

convertTMWorldBordersSimple(nameOnly = FALSE, simplify = TRUE)

Arguments

nameOnly
Logical specifying whether to only return the name without creating the file.

simplify
Logical specifying whether to create "_05", _02" and "_01" versions of the file that are simplified to 5%, 2% and 1%.

Value

Name of the dataset being created.

Details

A world borders shapefile is downloaded and converted to a SpatialPolygonsDataFrame with additional columns of data. The resulting file will be created in the package SpatialDataDir which is set with setSpatialDataDir().

This shapefile is a simplified version of the TMWorldBorders shapefile. Users may wish to use a higher resolution dataset when plotting.

References

http://thematicmapping.org/downloads/

See also