Convert Wikipedia timezone table to dataframe

Returns a dataframe version of the Wikipedia timezone table with the following columns:

timezone -- Olson timezone

UTC_offset -- hours between local timezone and UTC

UTC_DST_offset -- hours between local timezone daylight savings and UTC

countryCode -- ISO 3166-2 country code

longitude -- longitude of the Olson timezone city

latitude -- latitude of the Olson timezone city

status -- either 'Canonical', 'Alias' or 'Deprecated'

notes -- typically specifying the target of an 'Alias'

convertWikipediaTimezoneTable()

Value

Dataframe with 388 rows and 10 columns.

Details

Older named timezones from the table which are linked to more modern equivalents are not included in the returned dataframe.

References

https://en.wikipedia.org/wiki/List_of_tz_database_time_zones