Return country ISO codes at specified locations

Uses spatial comparison to determine which country polygons the locations fall into and returns the country code strings for those polygons.

If allData = TRUE, additional data is returned.

getCountryCode(
  lon,
  lat,
  dataset = "SimpleCountriesEEZ",
  countryCodes = NULL,
  allData = FALSE,
  useBuffering = FALSE
)

Arguments

lon
vector of longitudes in decimal degrees

lat
vector of latitudes in decimal degrees

dataset
name of spatial dataset to use

countryCodes
vector of countryCodes

allData
logical specifying whether a full dataframe should be returned

useBuffering
logical flag specifying the use of location buffering to find the nearest polygon if no target polygon is found

Value

Vector of ISO-3166-1 alpha-2 country codes.

References

http://www.naturalearthdata.com/downloads/10m-cultural-vectors/

See also

Examples

lon <- seq(0, 50)
lat <- seq(0, 50)
getCountryCode(lon, lat)

#>  [1] NA   NA   NA   "NG" "NG" "NG" "NG" "NG" "NG" "NG" "NG" "NG" "NG" "NG" "TD"
#> [16] "TD" "TD" "TD" "TD" "TD" "TD" "TD" "LY" "LY" "LY" "EG" "EG" "EG" "EG" "EG"
#> [31] "EG" "EG" "EG" "CY" "CY" "SY" "TR" "TR" "TR" "TR" "TR" "TR" "GE" "GE" "RU"
#> [46] "RU" "RU" "RU" "KZ" "KZ" "KZ"