List locally installed spatial datasets

Searches the directory set with setSpatialDataDir for locally installed spatial data and returns a list of dataset names that can be used with loadSpatialData.

If verbose = TRUE, a brief description is provided for each locally installed dataset.

installedSpatialData(verbose = TRUE)

Arguments

verbose
Logical specifying whether or not to print dataset descriptions.

Value

Vector of dataset names().